+++ include/GUI/Unit.hpp
@@ -5,14 +5,14 @@ enum GUIUnitType { Relative, Absolute };
 
 class GUIUnit {
 private:
-  float number;
+  double number;
   enum GUIUnitType type;
 
 public:
-  GUIUnit(float number = 0, GUIUnitType type = Relative);
+  GUIUnit(double number = 0, GUIUnitType type = Relative);
   bool is_absolute() const;
-  float to_pixels() const;
-  float to_pixels(float parent_length) const;
+  double to_pixels() const;
+  double to_pixels(double parent_length) const;
 };
 
 #endif
