+++ src/GUI.cpp
@@ -13,21 +13,21 @@ void GUI::add_container(GUIContainer *container) {
 
 void GUI::add_text(GUIText *text) { root_container->add_text(text); }
 
-void GUI::on_window_resize(float width, float height) {
+void GUI::on_window_resize(double width, double height) {
   root_container->on_resize(width, height);
 }
 
-void GUI::on_mouse_move(float x1, float y1, float x2, float y2) {
+void GUI::on_mouse_move(double x1, double y1, double x2, double y2) {
   root_container->on_mouse_move(x1, y1, x2, y2);
 }
 
-void GUI::on_mouse_down(float x, float y) {
+void GUI::on_mouse_down(double x, double y) {
   root_container->on_mouse_down(x, y);
 }
 
-void GUI::on_mouse_up(float x, float y) { root_container->on_mouse_up(x, y); }
+void GUI::on_mouse_up(double x, double y) { root_container->on_mouse_up(x, y); }
 
-void GUI::on_mouse_drag(float x1, float y1, float x2, float y2) {
+void GUI::on_mouse_drag(double x1, double y1, double x2, double y2) {
   root_container->on_mouse_drag(x1, y1, x2 - x1, y2 - y1);
   // root_container->update(root_container->get_width(), root_container->get_height());
 }
